recovery-agent/requirements.txt
langgraph
langchain-google-genai
confluent-kafka
python-dotenv
pydantic
langgraph-checkpoint-postgres
psycopg[binary]
